LayoutTests/platform/gtk/accessibility/custom-elements/textbox-expected.txt

This tests that aria fallback roles work correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL accessibilityController.accessibleElementById("multiline-textbox").role should be AXRole: AXTextArea. Was AXRole: AXTextField.
PASS accessibilityController.accessibleElementById("multiline-textbox").stringAttributeValue("AXPlaceholderValue") is "some text"
PASS accessibilityController.accessibleElementById("multiline-textbox").isAttributeSettable("AXValue") is false
PASS accessibilityController.accessibleElementById("singleline-textbox").role is "AXRole: AXTextField"
PASS accessibilityController.accessibleElementById("singleline-textbox").stringAttributeValue("AXPlaceholderValue") is "another text"
PASS accessibilityController.accessibleElementById("singleline-textbox").isAttributeSettable("AXValue") is true
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
